+++ CHANGELOG.md
@@ -20,6 +20,13 @@ Types of changes:
 ### Added
 - Coins which give you a little speed boost!
 - Screen shake camera effect
+- Expanded options menu with:
+  - Resolution
+  - Fullscreen
+  - Vsync
+  - High score reset
+  - Reset all options
+- Settings are now saved and loaded from a file
 - Command line options for starting the game or server:
   - --join \<ip:port\> - join a game at an ip address and optional port
   - --host - host a game as a client